// ListAppi/types/RecipeMeta.ts
export type MealType =
    | 'airfryer-ruoat'
    | 'keitot'
    | 'salaatit'
    | 'pastat'
    | 'hampurilaiset'
    | 'pihvit'
    | 'uuniruoat'
    | 'pataruoat'
    | 'tacot ja tortillat'
    | 'muu'

export type MainIngredient =
    | 'liha'
    | 'jauheliha'
    | 'makkara'
    | 'broileri'
    | 'kala'
    | 'äyriäiset'
    | 'kananmuna'
    | 'kasvis'
    | 'kasviproteiini'
    | 'muu'

export type DietType =
    | 'gluteeniton'
    | 'kananmunaton'
    | 'kasvis'
    | 'laktoositon'
    | 'maidoton'
    | 'vegaaninen'